requirements.txt
openai>=1.0.0
streamlit
watchdog
python-dotenv
